fix incorrect package name
