+++ platforms/folia/src/main/java/com/cjcrafter/scheduler/folia/FoliaGlobalScheduler.java
@@ -1,28 +1,26 @@
-package com.scheduler.folia;
+package com.cjcrafter.scheduler.folia;
 
-
-import com.cjcrafter.scheduler.AsyncSchedulerImplementation;
+import com.cjcrafter.scheduler.SchedulerImplementation;
 import com.cjcrafter.scheduler.TaskImplementation;
-import io.papermc.paper.threadedregions.scheduler.AsyncScheduler;
+import io.papermc.paper.threadedregions.scheduler.GlobalRegionScheduler;
 import io.papermc.paper.threadedregions.scheduler.ScheduledTask;
 import org.bukkit.plugin.Plugin;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
 @ApiStatus.Internal
-public class FoliaAsyncScheduler implements AsyncSchedulerImplementation {
+public class FoliaGlobalScheduler implements SchedulerImplementation {
 
     private final @NotNull Plugin plugin;
-    private final @NotNull AsyncScheduler asyncScheduler;
+    private final @NotNull GlobalRegionScheduler globalRegionScheduler;
 
     @ApiStatus.Internal
-    public FoliaAsyncScheduler(@NotNull Plugin plugin) {
+    public FoliaGlobalScheduler(@NotNull Plugin plugin) {
         this.plugin = plugin;
-        this.asyncScheduler = plugin.getServer().getAsyncScheduler();
+        this.globalRegionScheduler = plugin.getServer().getGlobalRegionScheduler();
     }
 
     private <T> @NotNull Consumer<ScheduledTask> buildFoliaConsumer(
@@ -37,28 +35,33 @@ public class FoliaAsyncScheduler implements AsyncSchedulerImplementation {
     }
 
     @Override
-    public <T> @NotNull TaskImplementation<T> runNow(@NotNull Function<TaskImplementation<T>, T> function) {
+    public void execute(@NotNull Runnable run) {
+        globalRegionScheduler.execute(plugin, run);
+    }
+
+    @Override
+    public @NotNull <T> TaskImplementation<T> run(@NotNull Function<TaskImplementation<T>, T> function) {
         FoliaTask<T> taskImplementation = new FoliaTask<>();
         Consumer<ScheduledTask> foliaConsumer = buildFoliaConsumer(taskImplementation, function);
-        ScheduledTask scheduledTask = asyncScheduler.runNow(plugin, foliaConsumer);
+        ScheduledTask scheduledTask = globalRegionScheduler.run(plugin, foliaConsumer);
         taskImplementation.setScheduledTask(scheduledTask);
         return taskImplementation;
     }
 
     @Override
-    public <T> @NotNull TaskImplementation<T> runDelayed(@NotNull Function<TaskImplementation<T>, T> function, long delay, @NotNull TimeUnit unit) {
+    public @NotNull <T> TaskImplementation<T> runDelayed(@NotNull Function<TaskImplementation<T>, T> function, long delay) {
         FoliaTask<T> taskImplementation = new FoliaTask<>();
         Consumer<ScheduledTask> foliaConsumer = buildFoliaConsumer(taskImplementation, function);
-        ScheduledTask scheduledTask = asyncScheduler.runDelayed(plugin, foliaConsumer, delay, unit);
+        ScheduledTask scheduledTask = globalRegionScheduler.runDelayed(plugin, foliaConsumer, delay);
         taskImplementation.setScheduledTask(scheduledTask);
         return taskImplementation;
     }
 
     @Override
-    public <T> @NotNull TaskImplementation<T> runAtFixedRate(@NotNull Function<TaskImplementation<T>, T> function, long delay, long period, @NotNull TimeUnit unit) {
+    public @NotNull <T> TaskImplementation<T> runAtFixedRate(@NotNull Function<TaskImplementation<T>, T> function, long delay, long period) {
         FoliaTask<T> taskImplementation = new FoliaTask<>();
         Consumer<ScheduledTask> foliaConsumer = buildFoliaConsumer(taskImplementation, function);
-        ScheduledTask scheduledTask = asyncScheduler.runAtFixedRate(plugin, foliaConsumer, delay, period, unit);
+        ScheduledTask scheduledTask = globalRegionScheduler.runAtFixedRate(plugin, foliaConsumer, delay, period);
         taskImplementation.setScheduledTask(scheduledTask);
         return taskImplementation;
     }
