+++ platforms/folia/src/main/java/com/cjcrafter/scheduler/folia/FoliaTask.java
@@ -1,4 +1,4 @@
-package com.scheduler.folia;
+package com.cjcrafter.scheduler.folia;
 
 import com.cjcrafter.scheduler.TaskImplementation;
 import io.papermc.paper.threadedregions.scheduler.ScheduledTask;
